+++ src/pdftl/utils/page_specs/help.py
@@ -17,8 +17,9 @@ def _help_topic_page_specs():
 
     1. Page range: Which pages to select.
 
-    2. Qualifiers and omissions: Filtering the selected pages by
-    parity (even/odd) and omitted ranges.
+    2. Qualifiers, step and omissions: Filtering the selected pages by
+    parity (even/odd), a step size to use when interpreting ranges,
+    and omitted ranges.
 
     3. Transformation modifiers: Applying rotation or scaling to
     the selected pages. This is ignored by some operations.
@@ -32,25 +33,25 @@ def _help_topic_page_specs():
 
     A page identifier can be:
 
-      an integer (e.g., `5`) representing that page (numbered
-      from page 1, the first page of the PDF file, regardless of
-      any page labelling),
+    * an integer (e.g., `5`) representing that page (numbered from
+      page 1, the first page of the PDF file, regardless of any page
+      labelling),
 
-      the keyword `end`,
+    * the keyword `end`,
 
-      or `r` followed by one of the two above types,
-      representing reverse numbering. So `r1` means the same as
-      `end`, and `rend` means the same as `1`.
+    * or `r` followed by one of the two above types, representing
+      reverse numbering. So `r1` means the same as `end`, and `rend`
+      means the same as `1`.
 
     The following page range formats are supported:
 
-    `<I>`: A single page identifier
+    * `<I>`: A single page identifier
 
-    `<I>-<J>`: A range of pages (e.g., `1-5`). If the start page
-    number is higher than the end page number (e.g., `5-1`),
-    then the pages are treated in reverse order.
+    * `<I>-<J>`: A range of pages (e.g., `1-5`). If the start page
+    number is higher than the end page number (e.g., `5-1`), then the
+    pages are treated in reverse order.
 
-    ### 2. Page qualifiers and omissions
+    ### 2. Page qualifiers, step and omissions
 
     #### Parity qualifiers
 
@@ -58,24 +59,45 @@ def _help_topic_page_specs():
     number. They are added immediately after the page
     range. Valid qualifiers are:
 
-    `even`: selects only even-numbered pages in the range (e.g.,
+    * `even`: selects only even-numbered pages in the range (e.g.,
     `1-10even`).
 
-    `odd`: selects only odd-numbered pages in the range (e.g.,
+    * `odd`: selects only odd-numbered pages in the range (e.g.,
     `odd` alone selects all odd pages).
+
+    ### Step size
+
+    An optional step size to use when interpreting page ranges,
+    starting from the first page in the range. The step size must an
+    integer, and at least 1. The default step size is 1.
+
+    Examples
+
+    * `step4` selects every 4th page (pages 1, 5, 9, ...) in the
+      document.
+
+    * `5-12step3` selects pages 5, 8, 11
+
+    * `12-5step3` selects pages 12, 9, 6
+
+    * `5-14step3` selects pages 5, 8, 11, 14
+
+    * `5-14step3even` selects pages 8, 11
+
+    You can also use `by` or `every` as alternatives to `step`.
 
     #### Omissions
 
     The `~` operator is used to exclude pages from the selection
     defined by the preceding page range and qualifiers.
 
-    `~<N>-<M>`: Omits a range of pages (e.g., `1-end~5-10` selects
+    * `~<N>-<M>`: Omits a range of pages (e.g., `1-end~5-10` selects
     all pages except 5 through 10).
 
-    `~<N>`: Omits a single page (e.g., `1-10~5` selects all pages
+    * `~<N>`: Omits a single page (e.g., `1-10~5` selects all pages
     from 1 to 10 except page 5).
 
-    `~r<N>`: Omits a single page counting backwards from the end
+    * `~r<N>`: Omits a single page counting backwards from the end
     (e.g., `~r1` omits the last page).
 
     ### 3. Transformation Modifiers
@@ -89,11 +111,11 @@ def _help_topic_page_specs():
     These modifiers adjust the page's current rotation property
     by adding or subtracting degrees.
 
-    right: Rotates 90 degrees clockwise (+90),
+    * `right`: Rotates 90 degrees clockwise (+90),
 
-    left: Rotates 90 degrees counter-clockwise (-90),
+    * `left`: Rotates 90 degrees counter-clockwise (-90),
 
-    down: Rotates 180 degrees (+180).
+    * `down`: Rotates 180 degrees (+180).
 
     #### Rotation (absolute)
 
@@ -101,22 +123,22 @@ def _help_topic_page_specs():
     to a fixed orientation (0, 90, 180, or 270 degrees) relative
     to the page's natural (unrotated) state.
 
-    `north`: Resets rotation to the natural page orientation,
+    * `north`: Resets rotation to the natural page orientation,
 
-    `east`: Sets rotation to 90 degrees clockwise,
+    * `east`: Sets rotation to 90 degrees clockwise,
 
-    `south`: Sets rotation to 180 degrees,
+    * `south`: Sets rotation to 180 degrees,
 
-    `west`: Sets rotation to 270 degrees clockwise or 90 degrees
+    * `west`: Sets rotation to 270 degrees clockwise or 90 degrees
     counter-clockwise.
 
     #### Scale and zoom
 
-    `x<N>`: Scales the page content by factor N. N is typically an
+    * `x<N>`: Scales the page content by factor N. N is typically an
     integer or decimal (e.g., `x2` doubles the size, `x0.5`
     halves it).
 
-    `z<N>`: Zoom in by N steps (or out if N is negative), where a
+    * `z<N>`: Zoom in by N steps (or out if N is negative), where a
     zoom of 1 step corresponds to enlarging A4 paper to A3. More
     technically, we scale by factor of 2^(N/2). (N can be any
     number). For example, z1 will scale A4 pages up to A3, and
@@ -124,7 +146,7 @@ def _help_topic_page_specs():
 
     #### Groups (Applies to all contents)
 
-    `[<Range>, <Range>]<Modifier>`: Applies the modifier to the entire
+    * `[<Range>, <Range>]<Modifier>`: Applies the modifier to the entire
     disjoint set of pages.
 
     Example: `[1-3, 5]x2` scales pages 1, 2, 3, and 5.
@@ -135,27 +157,27 @@ def _help_topic_page_specs():
 
     ### Examples
 
-    `1-5eastx2` selects pages 1 through 5, rotating them 90
+    * `1-5eastx2` selects pages 1 through 5, rotating them 90
     degrees clockwise (east) and scaling them by 2x.
 
 
-    `oddleftz-1` selects only the odd pages from the beginning
+    * `oddleftz-1` selects only the odd pages from the beginning
     to the end, rotating them 90 degrees counter-clockwise
     (left) and applying a zoom factor of z-1.
 
 
-    `1-end~3-5` or equivalently `~3-5` selects all pages except
+    * `1-end~3-5` or equivalently `~3-5` selects all pages except
     pages 3-5.
 
-    `~2downz1` selects all pages except page 2, rotating them by
+    * `~2downz1` selects all pages except page 2, rotating them by
     180 degrees and zooming in 1 step. This will likely need to
     be quoted to prevent your shell misinterpreting it. (The
     same goes for `~3-5`).
 
-    `end-r4` selects the last 4 pages, in reverse order.
+    * `end-r4` selects the last 4 pages, in reverse order.
 
-    `1,3,5` selects pages 1, 3, and 5.
+    * `1,3,5` selects pages 1, 3, and 5.
 
-    `[1,3,5]x2` selects pages 1, 3, and 5 and scales them all.
+    * `[1,3,5]x2` selects pages 1, 3, and 5 and scales them all.
 
     """